backend/requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.6
django-cleanup==9.0.0
django-cors-headers==4.8.0
django-import-export==4.3.9
django-nested-admin==4.1.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-monkey-business==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
whitenoise==6.10.0
